Some more naming consistency stuff
