+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/controls/ControlSingle.java
@@ -1,9 +1,12 @@
-package org.firstinspires.ftc.twenty403.Controls;
+package org.firstinspires.ftc.twenty403.controls;
 
 import org.firstinspires.ftc.twenty403.Robot;
-import org.firstinspires.ftc.twenty403.command.claw.ClawAutoCloseToggleCommand;
 import org.firstinspires.ftc.twenty403.command.claw.ClawCloseCommand;
 import org.firstinspires.ftc.twenty403.command.claw.ClawOpenCommand;
+import org.firstinspires.ftc.twenty403.command.drive.DriveCommand;
+import org.firstinspires.ftc.twenty403.command.drive.ResetGyroCommand;
+import org.firstinspires.ftc.twenty403.command.drive.SlowCommand;
+import org.firstinspires.ftc.twenty403.command.drive.TurboCommand;
 import org.firstinspires.ftc.twenty403.command.lift.LiftDownCommand;
 import org.firstinspires.ftc.twenty403.command.lift.LiftGroundJunctionCommand;
 import org.firstinspires.ftc.twenty403.command.lift.LiftHighJunctionCommand;
@@ -15,29 +18,33 @@ import org.firstinspires.ftc.twenty403.command.lift.LiftMoveUpOverrideCommand;
 import org.firstinspires.ftc.twenty403.command.lift.LiftSetZeroCommand;
 import org.firstinspires.ftc.twenty403.command.lift.LiftUpCommand;
 
+import com.technototes.library.command.CommandScheduler;
 import com.technototes.library.command.ConditionalCommand;
 import com.technototes.library.control.CommandButton;
 import com.technototes.library.control.CommandGamepad;
 import com.technototes.library.control.Stick;
 
-public class ControlOperator {
+public class ControlSingle {
 
     public Robot robot;
     public CommandGamepad gamepad;
 
     public Stick driveLeftStick, driveRightStick;
-    public CommandButton resetGyroButton, driveStraighten, turboButton;
-    public CommandButton liftDownButton, liftUpButton, clawOpenButton, clawCloseButton, liftHighOrOverrideZero;
-    public CommandButton liftGroundOrOverrideDown, liftLowOrOverrideUp, liftMediumOrToggleAutoClose, liftIntakePos;
+    public CommandButton resetGyroButton, driveStraight, turboButton;
+    public CommandButton liftDownButton, liftUpButton, clawOpenButton, clawCloseButton, liftOverrideZeroButton;
+    public CommandButton liftGroundOrOverrideDown, liftLowOrOverrideUp, liftMedium, liftHigh, liftIntakePos;
     public CommandButton override;
 
-    public ControlOperator(CommandGamepad g, Robot r) {
+    public ControlSingle(CommandGamepad g, Robot r) {
         this.robot = r;
         gamepad = g;
         override = g.leftTrigger.getAsButton(0.5);
 
         AssignNamedControllerButton();
 
+        if (Robot.RobotConstant.DRIVE_CONNECTED) {
+            bindDriveControls();
+        }
         if (Robot.RobotConstant.CLAW_CONNECTED) {
             bindClawControls();
         }
@@ -47,36 +54,54 @@ public class ControlOperator {
     }
 
     private void AssignNamedControllerButton() {
+        resetGyroButton = gamepad.rightStickButton;
+        driveLeftStick = gamepad.leftStick;
+        driveRightStick = gamepad.rightStick;
+        turboButton = gamepad.leftStickButton;
+        driveStraight = gamepad.rightTrigger.getAsButton(0.5);
+
         liftUpButton = gamepad.dpadRight;
+
         liftDownButton = gamepad.dpadDown;
         liftIntakePos = gamepad.dpadLeft;
 
         clawOpenButton = gamepad.rightBumper;
         clawCloseButton = gamepad.leftBumper;
 
-        liftMediumOrToggleAutoClose = gamepad.circle;
-        liftHighOrOverrideZero = gamepad.triangle;
+        liftMedium = gamepad.circle;
+        liftHigh = gamepad.triangle;
+
         liftGroundOrOverrideDown = gamepad.cross;
         liftLowOrOverrideUp = gamepad.square;
+        liftOverrideZeroButton = gamepad.triangle;
+
+        // TODO: Identify other controls for
+    }
+
+    public void bindDriveControls() {
+        CommandScheduler.getInstance()
+                .scheduleJoystick(
+                        new DriveCommand(robot.drivebaseSubsystem, driveLeftStick, driveRightStick, driveStraight));
+        turboButton.whenPressed(new TurboCommand(robot.drivebaseSubsystem));
+        turboButton.whenReleased(new SlowCommand(robot.drivebaseSubsystem));
+        // TODO: We probably want buttons to reset the Gyro...
+        resetGyroButton.whenPressed(new ResetGyroCommand(robot.drivebaseSubsystem));
     }
 
     public void bindClawControls() {
+        // TODO: Name & Bind claw controls
         clawOpenButton.whenPressed(new ClawOpenCommand(robot.clawSubsystem));
         clawCloseButton.whenReleased(new ClawCloseCommand(robot.clawSubsystem));
     }
 
     public void bindLiftControls() {
+        // TODO: Name & Bind lift controls
         liftUpButton.whenPressed(new LiftUpCommand(robot.liftSubsystem));
         liftDownButton.whenPressed(new LiftDownCommand(robot.liftSubsystem));
         liftIntakePos.whenPressed(new LiftIntakeCommand(robot.liftSubsystem));
-        liftHighOrOverrideZero.whenPressed(new ConditionalCommand(
-                override,
-                new LiftSetZeroCommand(robot.liftSubsystem),
-                new LiftHighJunctionCommand(robot.liftSubsystem)));
-        liftMediumOrToggleAutoClose.whenPressed(new ConditionalCommand(
-                override,
-                new ClawAutoCloseToggleCommand(robot.clawSubsystem),
-                new LiftMidJunctionCommand(robot.liftSubsystem)));
+        liftOverrideZeroButton.whenPressed(
+                new ConditionalCommand(override, new LiftSetZeroCommand(robot.liftSubsystem)));
+
         liftGroundOrOverrideDown.whenPressed(new ConditionalCommand(
                 override,
                 new LiftMoveDownOverrideCommand(robot.liftSubsystem),
@@ -85,5 +110,7 @@ public class ControlOperator {
                 override,
                 new LiftMoveUpOverrideCommand(robot.liftSubsystem),
                 new LiftLowJunctionCommand(robot.liftSubsystem)));
+        liftMedium.whenPressed(new LiftMidJunctionCommand(robot.liftSubsystem));
+        liftHigh.whenPressed(new LiftHighJunctionCommand(robot.liftSubsystem));
     }
 }
