+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/DualBlue.java
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.twenty403.opmode;
 
-import org.firstinspires.ftc.twenty403.Controls.ControlOperator;
-import org.firstinspires.ftc.twenty403.Controls.ControlsDriver;
+import org.firstinspires.ftc.twenty403.controls.ControlDriver;
+import org.firstinspires.ftc.twenty403.controls.ControlOperator;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
@@ -11,11 +11,11 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 
-@TeleOp(name = "Dual Blue TeleOp")
+@TeleOp(name = "Dual Blue")
 @SuppressWarnings("unused")
-public class DualBlueTeleOp extends CommandOpMode {
+public class DualBlue extends CommandOpMode {
     public Robot robot;
-    public ControlsDriver controlsDriver;
+    public ControlDriver controlsDriver;
     public ControlOperator controlsOperator;
     public Hardware hardware;
 
@@ -23,7 +23,7 @@ public class DualBlueTeleOp extends CommandOpMode {
     public void uponInit() {
         hardware = new Hardware(hardwareMap);
         robot = new Robot(hardware, Alliance.BLUE, StartingPosition.NEUTRAL);
-        controlsDriver = new ControlsDriver(driverGamepad, robot);
+        controlsDriver = new ControlDriver(driverGamepad, robot);
         controlsOperator = new ControlOperator(codriverGamepad, robot);
     }
 }
