+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/DualRed.java
@@ -1,29 +1,29 @@
 package org.firstinspires.ftc.twenty403.opmode;
 
-import org.firstinspires.ftc.twenty403.Controls.Controls;
+import org.firstinspires.ftc.twenty403.controls.ControlDriver;
+import org.firstinspires.ftc.twenty403.controls.ControlOperator;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
 
-import com.acmerobotics.dashboard.FtcDashboard;
-import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 
-@TeleOp(name = "Manual Testing (Single Controller)")
+@TeleOp(name = "Dual Red")
 @SuppressWarnings("unused")
-public class ManualTestingOpMode extends CommandOpMode {
+public class DualRed extends CommandOpMode {
     public Robot robot;
-    public Controls controls;
+    public ControlDriver controlsDriver;
+    public ControlOperator controlsOperator;
     public Hardware hardware;
 
     @Override
     public void uponInit() {
-        telemetry = new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry());
         hardware = new Hardware(hardwareMap);
-        robot = new Robot(hardware, Alliance.NONE, StartingPosition.NEUTRAL);
-        controls = new Controls(driverGamepad, robot);
+        robot = new Robot(hardware, Alliance.RED, StartingPosition.NEUTRAL);
+        controlsDriver = new ControlDriver(driverGamepad, robot);
+        controlsOperator = new ControlOperator(codriverGamepad, robot);
     }
 }
