+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/testing/AutoTester.java
@@ -1,6 +1,6 @@
-package org.firstinspires.ftc.twenty403.opmode;
+package org.firstinspires.ftc.twenty403.opmode.testing;
 
-import org.firstinspires.ftc.twenty403.Controls.Controls;
+import org.firstinspires.ftc.twenty403.controls.ControlSingle;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.VisionCommand;
@@ -16,11 +16,11 @@ import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 
 @Disabled
-@Autonomous(name = "Drivebase Test Auto")
+@Autonomous(name = "Auto Tester")
 @SuppressWarnings("unused")
-public class DrivebaseTestAutoOpMode extends CommandOpMode {
+public class AutoTester extends CommandOpMode {
     public Robot robot;
-    public Controls controls;
+    public ControlSingle controls;
     public Hardware hardware;
 
     @Override
