+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/testing/ManualTesting.java
@@ -1,10 +1,12 @@
-package org.firstinspires.ftc.twenty403.opmode;
+package org.firstinspires.ftc.twenty403.opmode.testing;
 
-import org.firstinspires.ftc.twenty403.Controls.Controls;
+import org.firstinspires.ftc.twenty403.controls.ControlSingle;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
 
+import com.acmerobotics.dashboard.FtcDashboard;
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
@@ -12,17 +14,18 @@ import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 
 @Disabled
-@TeleOp(name = "Black [No Alliance]")
+@TeleOp(name = "Manual Testing: 1 Controller")
 @SuppressWarnings("unused")
-public class BlackTeleOpMode extends CommandOpMode {
+public class ManualTesting extends CommandOpMode {
     public Robot robot;
-    public Controls controls;
+    public ControlSingle controls;
     public Hardware hardware;
 
     @Override
     public void uponInit() {
+        telemetry = new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry());
         hardware = new Hardware(hardwareMap);
         robot = new Robot(hardware, Alliance.NONE, StartingPosition.NEUTRAL);
-        controls = new Controls(driverGamepad, robot);
+        controls = new ControlSingle(driverGamepad, robot);
     }
 }
