+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/testing/RoadRunnerTest.java
@@ -1,6 +1,6 @@
-package org.firstinspires.ftc.twenty403.opmode;
+package org.firstinspires.ftc.twenty403.opmode.testing;
 
-import org.firstinspires.ftc.twenty403.Controls.Controls;
+import org.firstinspires.ftc.twenty403.controls.ControlSingle;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.autonomous.AutoConstants;
@@ -9,6 +9,7 @@ import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 import com.technototes.library.command.CommandScheduler;
 import com.technototes.library.command.SequentialCommandGroup;
@@ -16,11 +17,12 @@ import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 import com.technototes.path.command.TrajectorySequenceCommand;
 
-@Autonomous(name = "Straight Test")
+@Disabled
+@Autonomous(name = "Road Runner Test")
 @SuppressWarnings("unused")
-public class StraightTest extends CommandOpMode {
+public class RoadRunnerTest extends CommandOpMode {
     public Robot robot;
-    public Controls controls;
+    public ControlSingle controls;
     public Hardware hardware;
 
     @Override
